Use big integer instead integer

+++ database/migrations/2018_07_25_000000_create_love_reactant_reaction_counters_table.php
@@ -27,9 +27,9 @@ final class CreateLoveReactantReactionCountersTable extends Migration
         Schema::create('love_reactant_reaction_counters', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('reactant_id');
-            $table->unsignedInteger('reaction_type_id');
-            $table->unsignedInteger('count')->default(0);
-            $table->integer('weight')->default(0);
+            $table->unsignedBigInteger('reaction_type_id');
+            $table->unsignedBigInteger('count')->default(0);
+            $table->bigInteger('weight')->default(0);
             $table->timestamps();
 
             $table->index('reactant_id');
